+++ NEWS.md
@@ -5,10 +5,16 @@
 as a vector.
 * Cleared rownames on datastep exit.
 * Updated logo.
-* Added "MaxChar" column to dictionary.
+* Added "MaxChar" column to dictionary to hold the maximum number of characters
+for a data value in the column.  This is different from "width" which
+is the user-defined, proscribed width.  MaxChar is the actual width with no 
+padding.
 * Added "rda" engine to libname.
 * Allowed single value NSE on drop, keep, and by parameter of `datastep()`.
 For multiple values, use `v()` function.
+* Added `output()` function to datastep.
+* Improved datastep output column ordering. New variables were coming out
+in the opposite order they were added.  Now they are better.
 
 # libr 1.2.5
 * BREAKING CHANGE: Removed the `%eq%` operator from this package, as it has